remodeled add course page

+++ src/pages/Courses.jsx
@@ -70,43 +70,51 @@ export default function Courses() {
   };
 
   const handleDeleteCourse = (course) => {
-    fetch(`http://localhost:3000/api/courses/${course._id}`, {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data) {
-          alert(`${course.name} was deleted`);
-          window.location.reload(false);
-        } else {
-          alert("something went wrong");
-        }
-      });
+    let userConfirmed = confirm("Do you want to delete this course?");
+
+    if (userConfirmed) {
+      fetch(`http://localhost:3000/api/courses/${course._id}`, {
+        method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          if (data) {
+            alert(`${course.name} was deleted`);
+            window.location.reload(false);
+          } else {
+            alert("something went wrong");
+          }
+        });
+    } else {
+      return;
+    }
   };
 
   return (
     <section className=" h-screen mx-3 my-5 ">
       <div className="text-center ">
         {isAdmin === true && (
-          <Button onClick={handleAddCourse}>Add Course</Button>
+          <button className="m-2 px-12 py-2 bg-blue-500 rounded-md text-white font-bold w-34" onClick={handleAddCourse}>Add Course</button>
         )}
       </div>
       <div className="grid grid-cols-1 lg:grid-cols-3 md:grid-cols-2">
-        {isAdmin === true
-          ? courses.map((data, i) => (
-              <>
-                <div key={i} className="mx-5 my-5 lg:mt-12">
-                  <div className="mb-5 block  mx-5 p-6 bg-white border border-gray-200 rounded-lg shadow  dark:bg-gray-800 dark:border-gray-700 dark:hover:bg-gray-700">
-                    <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
-                      {data.name}
-                    </h5>
-                    <p className="font-normal text-gray-700  dark:text-gray-400">
-                      {data.description}
-                    </p>
-                    <div className="text-center flex flex-wrap items-center justify-center ">
+        {courses.map((data, i) => (
+          <>
+            <div key={data._id || i} className="mx-5 my-5 lg:mt-12">
+              <div  className="h-62 mb-5 block  mx-5 p-6 bg-white border border-gray-200 rounded-lg shadow  dark:bg-gray-800 dark:border-gray-700 dark:hover:bg-gray-700  text-center">
+                <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
+                  {data.name}
+                </h5>
+                <p className="font-normal text-gray-700  dark:text-gray-400">
+                  {data.description}
+                </p>
+                {isAdmin === false && <p className="font-normal text-gray-700  dark:text-gray-400">$ {data.price}</p>}
+                <div className="text-center flex flex-wrap items-center justify-center ">
+                  {isAdmin === true ? (
+                    <>
                       <button
                         className="m-2 px-12 py-2 bg-green-500 rounded-md text-white font-bold w-full"
                         onClick={editCoursePage}
@@ -114,55 +122,21 @@ export default function Courses() {
                         Edit Course
                       </button>
                       <button
-                        className="m-2 px-12 py-2 bg-red-500 rounded-md text-white font-bold  w-full"
-                        onClick={handleDeleteCourse}
+                        className="m-2 px-12 py-2 bg-red-500 rounded-md text-white font-bold w-full"
+                        onClick={() => handleDeleteCourse(data)}
                       >
                         Delete Course
                       </button>
-                    </div>
-                  </div>
+                    </>
+                  ) : (
+                    <Button onClick={handleEnrollCourse}>Enroll</Button>
+                  )}
                 </div>
-              </>
-            ))
-          : courses.map((data, i) => (
-              <>
-                <div key={i} className="mx-5 my-5 lg:mt-12">
-                  <div className="mb-5 block  mx-5 p-6 bg-white border border-gray-200 rounded-lg shadow  dark:bg-gray-800 dark:border-gray-700 dark:hover:bg-gray-700">
-                    <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
-                      {data.name}
-                    </h5>
-                    <p className="font-normal text-gray-700 dark:text-gray-400">
-                      {data.description}
-                    </p>
-                    <div>
-                      <Button onclick={handleEnrollCourse}>Enroll</Button>
-                    </div>
-                  </div>
-                </div>
-              </>
-            ))}
+              </div>
+            </div>
+          </>
+        ))}
       </div>
     </section>
   );
 }
-
-// {courses.map((data, i) => (
-//   <>
-//   <div key={i} className="mx-5 my-5 lg:mt-12   ">
-//     <div
-//       href="#"
-//       className="mb-5 block  mx-5 p-6 bg-white border border-gray-200 rounded-lg shadow  dark:bg-gray-800 dark:border-gray-700 dark:hover:bg-gray-700  "
-//     >
-//       <h5 className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
-//         {data.name}
-//       </h5>
-//       <p className="font-normal text-gray-700 dark:text-gray-400">
-//         {data.description}
-//       </p>
-//       <div className="">
-//         <Button onclick={handleEnrollCourse}>Enroll</Button>
-//       </div>
-//     </div>
-//   </div>
-// </>
-// ))}
